TS: use same command codes as MS3 do
MS3 (and MS2) does serial port to CAN commands translation.
It expects read/write/burn and etc command to have same codes
as it uses. RusEFI hitoricaly has different command codes.
So change it to align with MS devices.
This is first step for RE and MS coexistance on CAN bus.